rename package name 'specification' to 'customquery'
