+++ api-gate/src/main/java/com/llzw/apigate/web/controller/OrderController.java
@@ -4,8 +4,8 @@ import com.llzw.apigate.persistence.dao.AddressRepository;
 import com.llzw.apigate.persistence.dao.OrderRepository;
 import com.llzw.apigate.persistence.dao.ProductRepository;
 import com.llzw.apigate.persistence.dao.StockRepository;
-import com.llzw.apigate.persistence.dao.specification.SearchCriteriaSpecificationFactory;
-import com.llzw.apigate.persistence.dao.specification.SearchCriterion;
+import com.llzw.apigate.persistence.dao.customquery.SearchCriteriaSpecificationFactory;
+import com.llzw.apigate.persistence.dao.customquery.SearchCriterion;
 import com.llzw.apigate.persistence.entity.*;
 import com.llzw.apigate.web.dto.OrderDto;
 import com.llzw.apigate.web.util.StandardRestResponse;
